chore: move lockfile out of utils
