+++ src/utils.ts
@@ -1,4 +1,4 @@
-import type { EmojiVersion } from "./utils/lockfile";
+import type { EmojiVersion } from "./lockfile";
 import consola from "consola";
 import semver from "semver";
 import { NO_EMOJI_VERSIONS } from "./constants";
